First step to fixing the graph visualization
