+++ backend/src/routes/production-modules/module.service.ts
@@ -50,7 +50,6 @@ export class ModuleService {
             const moduleSkillDtos = await this.skillService.getSkillsOfModule(moduleDto.iri);
             moduleDto.skillDtos = moduleSkillDtos;
         }
-
         return productionModuleDtos;
     }
