+++ frontend/src/modules/graph-visualization/graph-visualization.routing.ts
@@ -3,18 +3,18 @@ import { NgModule } from '@angular/core';
 import {GraphVisualizationComponent} from './graph-visualization.component';
 
 const routes: Routes = [
-  { path: 'modules/:moduleName',
-  component: GraphVisualizationComponent,
-  
-   },
-   {path:'modules', component: GraphVisualizationComponent
+    { path: 'modules/:moduleName',
+        component: GraphVisualizationComponent,
 
-   }
+    },
+    {path:'modules', component: GraphVisualizationComponent
+
+    }
 ];
 
 @NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule]
 })
 
 export class GraphVisualizationRouter {}
